Struct aws_sdk_swf::client::fluent_builders::StartWorkflowExecution

pub struct StartWorkflowExecution<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> StartWorkflowExecution<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<StartWorkflowExecutionOutput, SdkError<StartWorkflowExecutionError>> where
R::Policy: SmithyRetryPolicy<StartWorkflowExecutionInputOperationOutputAlias, StartWorkflowExecutionOutput, StartWorkflowExecutionError, StartWorkflowExecutionInputOperationRetryAlias>,

pub fn domain(self, inp: impl Into<String>) -> Self

The name of the domain in which the workflow execution is created.

pub fn set_domain(self, input: Option<String>) -> Self

pub fn workflow_id(self, inp: impl Into<String>) -> Self

The user defined identifier associated with the workflow execution. You can use this to associate a custom identifier with the workflow execution. You may specify the same identifier if a workflow execution is logically a restart of a previous execution. You cannot have two open workflow executions with the same workflowId at the same time within the same domain.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f-\u009f). Also, it must not be the literal string arn.

pub fn set_workflow_id(self, input: Option<String>) -> Self

pub fn workflow_type(self, inp: WorkflowType) -> Self

The type of the workflow to start.

pub fn set_workflow_type(self, input: Option<WorkflowType>) -> Self

pub fn task_list(self, inp: TaskList) -> Self

The task list to use for the decision tasks generated for this workflow execution. This overrides the defaultTaskList specified when registering the workflow type.

A task list for this workflow execution must be specified either as a default for the workflow type or through this parameter. If neither this parameter is set nor a default task list was specified at registration time then a fault is returned.

The specified string must not start or end with whitespace. It must not contain a : (colon), / (slash), | (vertical bar), or any control characters (\u0000-\u001f | \u007f-\u009f). Also, it must not be the literal string arn.

pub fn set_task_list(self, input: Option<TaskList>) -> Self

pub fn task_priority(self, inp: impl Into<String>) -> Self

The task priority to use for this workflow execution. This overrides any default priority that was assigned when the workflow type was registered. If not set, then the default task priority for the workflow type is used. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide.

pub fn set_task_priority(self, input: Option<String>) -> Self

pub fn input(self, inp: impl Into<String>) -> Self

The input for the workflow execution. This is a free form string which should be meaningful to the workflow you are starting. This input is made available to the new workflow execution in the WorkflowExecutionStarted history event.

pub fn set_input(self, input: Option<String>) -> Self

pub fn execution_start_to_close_timeout(self, inp: impl Into<String>) -> Self

The total duration for this workflow execution. This overrides the defaultExecutionStartToCloseTimeout specified when registering the workflow type.

The duration is specified in seconds; an integer greater than or equal to 0. Exceeding this limit causes the workflow execution to time out. Unlike some of the other timeout parameters in Amazon SWF, you cannot specify a value of "NONE" for this timeout; there is a one-year max limit on the time that a workflow execution can run.

An execution start-to-close timeout must be specified either through this parameter or as a default when the workflow type is registered. If neither this parameter nor a default execution start-to-close timeout is specified, a fault is returned.

pub fn set_execution_start_to_close_timeout(self, input: Option<String>) -> Self

pub fn tag_list(self, inp: impl Into<String>) -> Self

Appends an item to tagList.

To override the contents of this collection use set_tag_list.

The list of tags to associate with the workflow execution. You can specify a maximum of 5 tags. You can list workflow executions with a specific tag by calling ListOpenWorkflowExecutions or ListClosedWorkflowExecutions and specifying a TagFilter.

pub fn set_tag_list(self, input: Option<Vec<String>>) -> Self

pub fn task_start_to_close_timeout(self, inp: impl Into<String>) -> Self

Specifies the maximum duration of decision tasks for this workflow execution. This parameter overrides the defaultTaskStartToCloseTimout specified when registering the workflow type using RegisterWorkflowType.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

A task start-to-close timeout for this workflow execution must be specified either as a default for the workflow type or through this parameter. If neither this parameter is set nor a default task start-to-close timeout was specified at registration time then a fault is returned.

pub fn set_task_start_to_close_timeout(self, input: Option<String>) -> Self

pub fn child_policy(self, inp: ChildPolicy) -> Self

If set, specifies the policy to use for the child workflow executions of this workflow execution if it is terminated, by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout. This policy overrides the default child policy specified when registering the workflow type using RegisterWorkflowType.

The supported child policies are:

TERMINATE – The child executions are terminated.

REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON – No action is taken. The child executions continue to run.

A child policy for this workflow execution must be specified either as a default for the workflow type or through this parameter. If neither this parameter is set nor a default child policy was specified at registration time then a fault is returned.

pub fn set_child_policy(self, input: Option<ChildPolicy>) -> Self

pub fn lambda_role(self, inp: impl Into<String>) -> Self

The IAM role to attach to this workflow execution.

Executions of this workflow type need IAM roles to invoke Lambda functions. If you don't attach an IAM role, any attempt to schedule a Lambda task fails. This results in a ScheduleLambdaFunctionFailed history event. For more information, see https://docs.aws.amazon.com/amazonswf/latest/developerguide/lambda-task.html in the Amazon SWF Developer Guide.

pub fn set_lambda_role(self, input: Option<String>) -> Self