Module aws_sdk_swf::error

Modules

default_undefined_fault

See DefaultUndefinedFault

domain_already_exists_fault

See DomainAlreadyExistsFault

domain_deprecated_fault

See DomainDeprecatedFault

limit_exceeded_fault

See LimitExceededFault

operation_not_permitted_fault

See OperationNotPermittedFault

too_many_tags_fault

See TooManyTagsFault

type_already_exists_fault

See TypeAlreadyExistsFault

type_deprecated_fault

See TypeDeprecatedFault

unknown_resource_fault

See UnknownResourceFault

workflow_execution_already_started_fault

See WorkflowExecutionAlreadyStartedFault

Structs

CountClosedWorkflowExecutionsError

CountOpenWorkflowExecutionsError

CountPendingActivityTasksError

CountPendingDecisionTasksError

DefaultUndefinedFault

The StartWorkflowExecution API action was called without the required

DeprecateActivityTypeError

DeprecateDomainError

DeprecateWorkflowTypeError

DescribeActivityTypeError

DescribeDomainError

DescribeWorkflowExecutionError

DescribeWorkflowTypeError

DomainAlreadyExistsFault

Returned if the domain already exists. You may get this fault if you are registering a domain that is either already registered or deprecated, or if you undeprecate a domain that is currently registered.

DomainDeprecatedFault

Returned when the specified domain has been deprecated.

GetWorkflowExecutionHistoryError

LimitExceededFault

Returned by any operation if a system imposed limitation has been reached. To address this fault you should either clean up unused resources or increase the limit by contacting AWS.

ListActivityTypesError

ListClosedWorkflowExecutionsError

ListDomainsError

ListOpenWorkflowExecutionsError

ListTagsForResourceError

ListWorkflowTypesError

OperationNotPermittedFault

Returned when the caller doesn't have sufficient permissions to invoke the action.

PollForActivityTaskError

PollForDecisionTaskError

RecordActivityTaskHeartbeatError

RegisterActivityTypeError

RegisterDomainError

RegisterWorkflowTypeError

RequestCancelWorkflowExecutionError

RespondActivityTaskCanceledError

RespondActivityTaskCompletedError

RespondActivityTaskFailedError

RespondDecisionTaskCompletedError

SignalWorkflowExecutionError

StartWorkflowExecutionError

TagResourceError

TerminateWorkflowExecutionError

TooManyTagsFault

You've exceeded the number of tags allowed for a domain.

TypeAlreadyExistsFault

Returned if the type already exists in the specified domain. You may get this fault if you are registering a type that is either already registered or deprecated, or if you undeprecate a type that is currently registered.

TypeDeprecatedFault

Returned when the specified activity or workflow type was already deprecated.

UndeprecateActivityTypeError

UndeprecateDomainError

UndeprecateWorkflowTypeError

UnknownResourceFault

Returned when the named resource cannot be found with in the scope of this operation (region or domain). This could happen if the named resource was never created or is no longer available for this operation.

UntagResourceError

WorkflowExecutionAlreadyStartedFault

Returned by StartWorkflowExecution when an open execution with the same workflowId is already running in

Enums

CountClosedWorkflowExecutionsErrorKind

CountOpenWorkflowExecutionsErrorKind

CountPendingActivityTasksErrorKind

CountPendingDecisionTasksErrorKind

DeprecateActivityTypeErrorKind

DeprecateDomainErrorKind

DeprecateWorkflowTypeErrorKind

DescribeActivityTypeErrorKind

DescribeDomainErrorKind

DescribeWorkflowExecutionErrorKind

DescribeWorkflowTypeErrorKind

GetWorkflowExecutionHistoryErrorKind

ListActivityTypesErrorKind

ListClosedWorkflowExecutionsErrorKind

ListDomainsErrorKind

ListOpenWorkflowExecutionsErrorKind

ListTagsForResourceErrorKind

ListWorkflowTypesErrorKind

PollForActivityTaskErrorKind

PollForDecisionTaskErrorKind

RecordActivityTaskHeartbeatErrorKind

RegisterActivityTypeErrorKind

RegisterDomainErrorKind

RegisterWorkflowTypeErrorKind

RequestCancelWorkflowExecutionErrorKind

RespondActivityTaskCanceledErrorKind

RespondActivityTaskCompletedErrorKind

RespondActivityTaskFailedErrorKind

RespondDecisionTaskCompletedErrorKind

SignalWorkflowExecutionErrorKind

StartWorkflowExecutionErrorKind

TagResourceErrorKind

TerminateWorkflowExecutionErrorKind

UndeprecateActivityTypeErrorKind

UndeprecateDomainErrorKind

UndeprecateWorkflowTypeErrorKind

UntagResourceErrorKind