Module aws_sdk_swf::model

Modules

activity_task_cancel_requested_event_attributes

See ActivityTaskCancelRequestedEventAttributes

activity_task_canceled_event_attributes

See ActivityTaskCanceledEventAttributes

activity_task_completed_event_attributes

See ActivityTaskCompletedEventAttributes

activity_task_failed_event_attributes

See ActivityTaskFailedEventAttributes

activity_task_scheduled_event_attributes

See ActivityTaskScheduledEventAttributes

activity_task_started_event_attributes

See ActivityTaskStartedEventAttributes

activity_task_timed_out_event_attributes

See ActivityTaskTimedOutEventAttributes

activity_type

See ActivityType

activity_type_configuration

See ActivityTypeConfiguration

activity_type_info

See ActivityTypeInfo

cancel_timer_decision_attributes

See CancelTimerDecisionAttributes

cancel_timer_failed_event_attributes

See CancelTimerFailedEventAttributes

cancel_workflow_execution_decision_attributes

See CancelWorkflowExecutionDecisionAttributes

cancel_workflow_execution_failed_event_attributes

See CancelWorkflowExecutionFailedEventAttributes

child_workflow_execution_canceled_event_attributes

See ChildWorkflowExecutionCanceledEventAttributes

child_workflow_execution_completed_event_attributes

See ChildWorkflowExecutionCompletedEventAttributes

child_workflow_execution_failed_event_attributes

See ChildWorkflowExecutionFailedEventAttributes

child_workflow_execution_started_event_attributes

See ChildWorkflowExecutionStartedEventAttributes

child_workflow_execution_terminated_event_attributes

See ChildWorkflowExecutionTerminatedEventAttributes

child_workflow_execution_timed_out_event_attributes

See ChildWorkflowExecutionTimedOutEventAttributes

close_status_filter

See CloseStatusFilter

complete_workflow_execution_decision_attributes

See CompleteWorkflowExecutionDecisionAttributes

complete_workflow_execution_failed_event_attributes

See CompleteWorkflowExecutionFailedEventAttributes

continue_as_new_workflow_execution_decision_attributes

See ContinueAsNewWorkflowExecutionDecisionAttributes

continue_as_new_workflow_execution_failed_event_attributes

See ContinueAsNewWorkflowExecutionFailedEventAttributes

decision

See Decision

decision_task_completed_event_attributes

See DecisionTaskCompletedEventAttributes

decision_task_scheduled_event_attributes

See DecisionTaskScheduledEventAttributes

decision_task_started_event_attributes

See DecisionTaskStartedEventAttributes

decision_task_timed_out_event_attributes

See DecisionTaskTimedOutEventAttributes

domain_configuration

See DomainConfiguration

domain_info

See DomainInfo

execution_time_filter

See ExecutionTimeFilter

external_workflow_execution_cancel_requested_event_attributes

See ExternalWorkflowExecutionCancelRequestedEventAttributes

external_workflow_execution_signaled_event_attributes

See ExternalWorkflowExecutionSignaledEventAttributes

fail_workflow_execution_decision_attributes

See FailWorkflowExecutionDecisionAttributes

fail_workflow_execution_failed_event_attributes

See FailWorkflowExecutionFailedEventAttributes

history_event

See HistoryEvent

lambda_function_completed_event_attributes

See LambdaFunctionCompletedEventAttributes

lambda_function_failed_event_attributes

See LambdaFunctionFailedEventAttributes

lambda_function_scheduled_event_attributes

See LambdaFunctionScheduledEventAttributes

lambda_function_started_event_attributes

See LambdaFunctionStartedEventAttributes

lambda_function_timed_out_event_attributes

See LambdaFunctionTimedOutEventAttributes

marker_recorded_event_attributes

See MarkerRecordedEventAttributes

record_marker_decision_attributes

See RecordMarkerDecisionAttributes

record_marker_failed_event_attributes

See RecordMarkerFailedEventAttributes

request_cancel_activity_task_decision_attributes

See RequestCancelActivityTaskDecisionAttributes

request_cancel_activity_task_failed_event_attributes

See RequestCancelActivityTaskFailedEventAttributes

request_cancel_external_workflow_execution_decision_attributes

See RequestCancelExternalWorkflowExecutionDecisionAttributes

request_cancel_external_workflow_execution_failed_event_attributes

See RequestCancelExternalWorkflowExecutionFailedEventAttributes

request_cancel_external_workflow_execution_initiated_event_attributes

See RequestCancelExternalWorkflowExecutionInitiatedEventAttributes

resource_tag

See ResourceTag

schedule_activity_task_decision_attributes

See ScheduleActivityTaskDecisionAttributes

schedule_activity_task_failed_event_attributes

See ScheduleActivityTaskFailedEventAttributes

schedule_lambda_function_decision_attributes

See ScheduleLambdaFunctionDecisionAttributes

schedule_lambda_function_failed_event_attributes

See ScheduleLambdaFunctionFailedEventAttributes

signal_external_workflow_execution_decision_attributes

See SignalExternalWorkflowExecutionDecisionAttributes

signal_external_workflow_execution_failed_event_attributes

See SignalExternalWorkflowExecutionFailedEventAttributes

signal_external_workflow_execution_initiated_event_attributes

See SignalExternalWorkflowExecutionInitiatedEventAttributes

start_child_workflow_execution_decision_attributes

See StartChildWorkflowExecutionDecisionAttributes

start_child_workflow_execution_failed_event_attributes

See StartChildWorkflowExecutionFailedEventAttributes

start_child_workflow_execution_initiated_event_attributes

See StartChildWorkflowExecutionInitiatedEventAttributes

start_lambda_function_failed_event_attributes

See StartLambdaFunctionFailedEventAttributes

start_timer_decision_attributes

See StartTimerDecisionAttributes

start_timer_failed_event_attributes

See StartTimerFailedEventAttributes

tag_filter

See TagFilter

task_list

See TaskList

timer_canceled_event_attributes

See TimerCanceledEventAttributes

timer_fired_event_attributes

See TimerFiredEventAttributes

timer_started_event_attributes

See TimerStartedEventAttributes

workflow_execution

See WorkflowExecution

workflow_execution_cancel_requested_event_attributes

See WorkflowExecutionCancelRequestedEventAttributes

workflow_execution_canceled_event_attributes

See WorkflowExecutionCanceledEventAttributes

workflow_execution_completed_event_attributes

See WorkflowExecutionCompletedEventAttributes

workflow_execution_configuration

See WorkflowExecutionConfiguration

workflow_execution_continued_as_new_event_attributes

See WorkflowExecutionContinuedAsNewEventAttributes

workflow_execution_failed_event_attributes

See WorkflowExecutionFailedEventAttributes

workflow_execution_filter

See WorkflowExecutionFilter

workflow_execution_info

See WorkflowExecutionInfo

workflow_execution_open_counts

See WorkflowExecutionOpenCounts

workflow_execution_signaled_event_attributes

See WorkflowExecutionSignaledEventAttributes

workflow_execution_started_event_attributes

See WorkflowExecutionStartedEventAttributes

workflow_execution_terminated_event_attributes

See WorkflowExecutionTerminatedEventAttributes

workflow_execution_timed_out_event_attributes

See WorkflowExecutionTimedOutEventAttributes

workflow_type

See WorkflowType

workflow_type_configuration

See WorkflowTypeConfiguration

workflow_type_filter

See WorkflowTypeFilter

workflow_type_info

See WorkflowTypeInfo

Structs

ActivityTaskCancelRequestedEventAttributes

Provides the details of the ActivityTaskCancelRequested event.

ActivityTaskCanceledEventAttributes

Provides the details of the ActivityTaskCanceled event.

ActivityTaskCompletedEventAttributes

Provides the details of the ActivityTaskCompleted event.

ActivityTaskFailedEventAttributes

Provides the details of the ActivityTaskFailed event.

ActivityTaskScheduledEventAttributes

Provides the details of the ActivityTaskScheduled event.

ActivityTaskStartedEventAttributes

Provides the details of the ActivityTaskStarted event.

ActivityTaskTimedOutEventAttributes

Provides the details of the ActivityTaskTimedOut event.

ActivityType

Represents an activity type.

ActivityTypeConfiguration

Configuration settings registered with the activity type.

ActivityTypeInfo

Detailed information about an activity type.

CancelTimerDecisionAttributes

Provides the details of the CancelTimer decision.

CancelTimerFailedEventAttributes

Provides the details of the CancelTimerFailed event.

CancelWorkflowExecutionDecisionAttributes

Provides the details of the CancelWorkflowExecution decision.

CancelWorkflowExecutionFailedEventAttributes

Provides the details of the CancelWorkflowExecutionFailed event.

ChildWorkflowExecutionCanceledEventAttributes

Provide details of the ChildWorkflowExecutionCanceled event.

ChildWorkflowExecutionCompletedEventAttributes

Provides the details of the ChildWorkflowExecutionCompleted event.

ChildWorkflowExecutionFailedEventAttributes

Provides the details of the ChildWorkflowExecutionFailed event.

ChildWorkflowExecutionStartedEventAttributes

Provides the details of the ChildWorkflowExecutionStarted event.

ChildWorkflowExecutionTerminatedEventAttributes

Provides the details of the ChildWorkflowExecutionTerminated event.

ChildWorkflowExecutionTimedOutEventAttributes

Provides the details of the ChildWorkflowExecutionTimedOut event.

CloseStatusFilter

Used to filter the closed workflow executions in visibility APIs by their close status.

CompleteWorkflowExecutionDecisionAttributes

Provides the details of the CompleteWorkflowExecution decision.

CompleteWorkflowExecutionFailedEventAttributes

Provides the details of the CompleteWorkflowExecutionFailed event.

ContinueAsNewWorkflowExecutionDecisionAttributes

Provides the details of the ContinueAsNewWorkflowExecution decision.

ContinueAsNewWorkflowExecutionFailedEventAttributes

Provides the details of the ContinueAsNewWorkflowExecutionFailed event.

Decision

Specifies a decision made by the decider. A decision can be one of these types:

DecisionTaskCompletedEventAttributes

Provides the details of the DecisionTaskCompleted event.

DecisionTaskScheduledEventAttributes

Provides details about the DecisionTaskScheduled event.

DecisionTaskStartedEventAttributes

Provides the details of the DecisionTaskStarted event.

DecisionTaskTimedOutEventAttributes

Provides the details of the DecisionTaskTimedOut event.

DomainConfiguration

Contains the configuration settings of a domain.

DomainInfo

Contains general information about a domain.

ExecutionTimeFilter

Used to filter the workflow executions in visibility APIs by various time-based rules. Each parameter, if

ExternalWorkflowExecutionCancelRequestedEventAttributes

Provides the details of the ExternalWorkflowExecutionCancelRequested event.

ExternalWorkflowExecutionSignaledEventAttributes

Provides the details of the ExternalWorkflowExecutionSignaled event.

FailWorkflowExecutionDecisionAttributes

Provides the details of the FailWorkflowExecution decision.

FailWorkflowExecutionFailedEventAttributes

Provides the details of the FailWorkflowExecutionFailed event.

HistoryEvent

Event within a workflow execution. A history event can be one of these types:

LambdaFunctionCompletedEventAttributes

Provides the details of the LambdaFunctionCompleted event. It isn't set

LambdaFunctionFailedEventAttributes

Provides the details of the LambdaFunctionFailed event. It isn't set for

LambdaFunctionScheduledEventAttributes

Provides the details of the LambdaFunctionScheduled event. It isn't set

LambdaFunctionStartedEventAttributes

Provides the details of the LambdaFunctionStarted event. It isn't set for

LambdaFunctionTimedOutEventAttributes

Provides details of the LambdaFunctionTimedOut event.

MarkerRecordedEventAttributes

Provides the details of the MarkerRecorded event.

RecordMarkerDecisionAttributes

Provides the details of the RecordMarker decision.

RecordMarkerFailedEventAttributes

Provides the details of the RecordMarkerFailed event.

RequestCancelActivityTaskDecisionAttributes

Provides the details of the RequestCancelActivityTask decision.

RequestCancelActivityTaskFailedEventAttributes

Provides the details of the RequestCancelActivityTaskFailed event.

RequestCancelExternalWorkflowExecutionDecisionAttributes

Provides the details of the RequestCancelExternalWorkflowExecution decision.

RequestCancelExternalWorkflowExecutionFailedEventAttributes

Provides the details of the RequestCancelExternalWorkflowExecutionFailed event.

RequestCancelExternalWorkflowExecutionInitiatedEventAttributes

Provides the details of the RequestCancelExternalWorkflowExecutionInitiated event.

ResourceTag

Tags are key-value pairs that can be associated with Amazon SWF state machines and

ScheduleActivityTaskDecisionAttributes

Provides the details of the ScheduleActivityTask decision.

ScheduleActivityTaskFailedEventAttributes

Provides the details of the ScheduleActivityTaskFailed event.

ScheduleLambdaFunctionDecisionAttributes

Decision attributes specified in scheduleLambdaFunctionDecisionAttributes within the list of

ScheduleLambdaFunctionFailedEventAttributes

Provides the details of the ScheduleLambdaFunctionFailed event. It isn't

SignalExternalWorkflowExecutionDecisionAttributes

Provides the details of the SignalExternalWorkflowExecution decision.

SignalExternalWorkflowExecutionFailedEventAttributes

Provides the details of the SignalExternalWorkflowExecutionFailed event.

SignalExternalWorkflowExecutionInitiatedEventAttributes

Provides the details of the SignalExternalWorkflowExecutionInitiated event.

StartChildWorkflowExecutionDecisionAttributes

Provides the details of the StartChildWorkflowExecution decision.

StartChildWorkflowExecutionFailedEventAttributes

Provides the details of the StartChildWorkflowExecutionFailed event.

StartChildWorkflowExecutionInitiatedEventAttributes

Provides the details of the StartChildWorkflowExecutionInitiated event.

StartLambdaFunctionFailedEventAttributes

Provides the details of the StartLambdaFunctionFailed event. It isn't set

StartTimerDecisionAttributes

Provides the details of the StartTimer decision.

StartTimerFailedEventAttributes

Provides the details of the StartTimerFailed event.

TagFilter

Used to filter the workflow executions in visibility APIs based on a tag.

TaskList

Represents a task list.

TimerCanceledEventAttributes

TimerFiredEventAttributes

Provides the details of the TimerFired event.

TimerStartedEventAttributes

Provides the details of the TimerStarted event.

WorkflowExecution

Represents a workflow execution.

WorkflowExecutionCancelRequestedEventAttributes

Provides the details of the WorkflowExecutionCancelRequested event.

WorkflowExecutionCanceledEventAttributes

Provides the details of the WorkflowExecutionCanceled event.

WorkflowExecutionCompletedEventAttributes

Provides the details of the WorkflowExecutionCompleted event.

WorkflowExecutionConfiguration

The configuration settings for a workflow execution including timeout values, tasklist etc. These configuration settings are determined from the defaults specified when registering the workflow type and those specified when starting the workflow execution.

WorkflowExecutionContinuedAsNewEventAttributes

Provides the details of the WorkflowExecutionContinuedAsNew event.

WorkflowExecutionFailedEventAttributes

Provides the details of the WorkflowExecutionFailed event.

WorkflowExecutionFilter

Used to filter the workflow executions in visibility APIs by their workflowId.

WorkflowExecutionInfo

Contains information about a workflow execution.

WorkflowExecutionOpenCounts

Contains the counts of open tasks, child workflow executions and timers for a workflow execution.

WorkflowExecutionSignaledEventAttributes

Provides the details of the WorkflowExecutionSignaled event.

WorkflowExecutionStartedEventAttributes

Provides details of WorkflowExecutionStarted event.

WorkflowExecutionTerminatedEventAttributes

Provides the details of the WorkflowExecutionTerminated event.

WorkflowExecutionTimedOutEventAttributes

Provides the details of the WorkflowExecutionTimedOut event.

WorkflowType

Represents a workflow type.

WorkflowTypeConfiguration

The configuration settings of a workflow type.

WorkflowTypeFilter

Used to filter workflow execution query results by type. Each parameter, if specified, defines a rule that must be satisfied by each returned result.

WorkflowTypeInfo

Contains information about a workflow type.

Enums

ActivityTaskTimeoutType

CancelTimerFailedCause

CancelWorkflowExecutionFailedCause

ChildPolicy

CloseStatus

CompleteWorkflowExecutionFailedCause

ContinueAsNewWorkflowExecutionFailedCause

DecisionTaskTimeoutType

DecisionType

EventType

ExecutionStatus

FailWorkflowExecutionFailedCause

LambdaFunctionTimeoutType

RecordMarkerFailedCause

RegistrationStatus

RequestCancelActivityTaskFailedCause

RequestCancelExternalWorkflowExecutionFailedCause

ScheduleActivityTaskFailedCause

ScheduleLambdaFunctionFailedCause

SignalExternalWorkflowExecutionFailedCause

StartChildWorkflowExecutionFailedCause

StartLambdaFunctionFailedCause

StartTimerFailedCause

WorkflowExecutionCancelRequestedCause

WorkflowExecutionTerminatedCause

WorkflowExecutionTimeoutType