Struct aws_sdk_swf::model::history_event::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for HistoryEvent

Implementations

impl Builder

pub fn event_timestamp(self, input: Instant) -> Self

The date and time when the event occurred.

pub fn set_event_timestamp(self, input: Option<Instant>) -> Self

pub fn event_type(self, input: EventType) -> Self

The type of the history event.

pub fn set_event_type(self, input: Option<EventType>) -> Self

pub fn event_id(self, input: i64) -> Self

The system generated ID of the event. This ID uniquely identifies the event with in the workflow execution history.

pub fn set_event_id(self, input: Option<i64>) -> Self

pub fn workflow_execution_started_event_attributes(
self,
input: WorkflowExecutionStartedEventAttributes
) -> Self

If the event is of type WorkflowExecutionStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_workflow_execution_started_event_attributes(
self,
input: Option<WorkflowExecutionStartedEventAttributes>
) -> Self

pub fn workflow_execution_completed_event_attributes(
self,
input: WorkflowExecutionCompletedEventAttributes
) -> Self

If the event is of type WorkflowExecutionCompleted then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_workflow_execution_completed_event_attributes(
self,
input: Option<WorkflowExecutionCompletedEventAttributes>
) -> Self

pub fn complete_workflow_execution_failed_event_attributes(
self,
input: CompleteWorkflowExecutionFailedEventAttributes
) -> Self

If the event is of type CompleteWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_complete_workflow_execution_failed_event_attributes(
self,
input: Option<CompleteWorkflowExecutionFailedEventAttributes>
) -> Self

pub fn workflow_execution_failed_event_attributes(
self,
input: WorkflowExecutionFailedEventAttributes
) -> Self

If the event is of type WorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_workflow_execution_failed_event_attributes(
self,
input: Option<WorkflowExecutionFailedEventAttributes>
) -> Self

pub fn fail_workflow_execution_failed_event_attributes(
self,
input: FailWorkflowExecutionFailedEventAttributes
) -> Self

If the event is of type FailWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_fail_workflow_execution_failed_event_attributes(
self,
input: Option<FailWorkflowExecutionFailedEventAttributes>
) -> Self

pub fn workflow_execution_timed_out_event_attributes(
self,
input: WorkflowExecutionTimedOutEventAttributes
) -> Self

If the event is of type WorkflowExecutionTimedOut then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_workflow_execution_timed_out_event_attributes(
self,
input: Option<WorkflowExecutionTimedOutEventAttributes>
) -> Self

pub fn workflow_execution_canceled_event_attributes(
self,
input: WorkflowExecutionCanceledEventAttributes
) -> Self

If the event is of type WorkflowExecutionCanceled then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_workflow_execution_canceled_event_attributes(
self,
input: Option<WorkflowExecutionCanceledEventAttributes>
) -> Self

pub fn cancel_workflow_execution_failed_event_attributes(
self,
input: CancelWorkflowExecutionFailedEventAttributes
) -> Self

If the event is of type CancelWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_cancel_workflow_execution_failed_event_attributes(
self,
input: Option<CancelWorkflowExecutionFailedEventAttributes>
) -> Self

pub fn workflow_execution_continued_as_new_event_attributes(
self,
input: WorkflowExecutionContinuedAsNewEventAttributes
) -> Self

If the event is of type WorkflowExecutionContinuedAsNew then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_workflow_execution_continued_as_new_event_attributes(
self,
input: Option<WorkflowExecutionContinuedAsNewEventAttributes>
) -> Self

pub fn continue_as_new_workflow_execution_failed_event_attributes(
self,
input: ContinueAsNewWorkflowExecutionFailedEventAttributes
) -> Self

If the event is of type ContinueAsNewWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_continue_as_new_workflow_execution_failed_event_attributes(
self,
input: Option<ContinueAsNewWorkflowExecutionFailedEventAttributes>
) -> Self

pub fn workflow_execution_terminated_event_attributes(
self,
input: WorkflowExecutionTerminatedEventAttributes
) -> Self

If the event is of type WorkflowExecutionTerminated then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_workflow_execution_terminated_event_attributes(
self,
input: Option<WorkflowExecutionTerminatedEventAttributes>
) -> Self

pub fn workflow_execution_cancel_requested_event_attributes(
self,
input: WorkflowExecutionCancelRequestedEventAttributes
) -> Self

If the event is of type WorkflowExecutionCancelRequested then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_workflow_execution_cancel_requested_event_attributes(
self,
input: Option<WorkflowExecutionCancelRequestedEventAttributes>
) -> Self

pub fn decision_task_scheduled_event_attributes(
self,
input: DecisionTaskScheduledEventAttributes
) -> Self

If the event is of type DecisionTaskScheduled then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_decision_task_scheduled_event_attributes(
self,
input: Option<DecisionTaskScheduledEventAttributes>
) -> Self

pub fn decision_task_started_event_attributes(
self,
input: DecisionTaskStartedEventAttributes
) -> Self

If the event is of type DecisionTaskStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_decision_task_started_event_attributes(
self,
input: Option<DecisionTaskStartedEventAttributes>
) -> Self

pub fn decision_task_completed_event_attributes(
self,
input: DecisionTaskCompletedEventAttributes
) -> Self

If the event is of type DecisionTaskCompleted then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_decision_task_completed_event_attributes(
self,
input: Option<DecisionTaskCompletedEventAttributes>
) -> Self

pub fn decision_task_timed_out_event_attributes(
self,
input: DecisionTaskTimedOutEventAttributes
) -> Self

If the event is of type DecisionTaskTimedOut then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_decision_task_timed_out_event_attributes(
self,
input: Option<DecisionTaskTimedOutEventAttributes>
) -> Self

pub fn activity_task_scheduled_event_attributes(
self,
input: ActivityTaskScheduledEventAttributes
) -> Self

If the event is of type ActivityTaskScheduled then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_activity_task_scheduled_event_attributes(
self,
input: Option<ActivityTaskScheduledEventAttributes>
) -> Self

pub fn activity_task_started_event_attributes(
self,
input: ActivityTaskStartedEventAttributes
) -> Self

If the event is of type ActivityTaskStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_activity_task_started_event_attributes(
self,
input: Option<ActivityTaskStartedEventAttributes>
) -> Self

pub fn activity_task_completed_event_attributes(
self,
input: ActivityTaskCompletedEventAttributes
) -> Self

If the event is of type ActivityTaskCompleted then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_activity_task_completed_event_attributes(
self,
input: Option<ActivityTaskCompletedEventAttributes>
) -> Self

pub fn activity_task_failed_event_attributes(
self,
input: ActivityTaskFailedEventAttributes
) -> Self

If the event is of type ActivityTaskFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_activity_task_failed_event_attributes(
self,
input: Option<ActivityTaskFailedEventAttributes>
) -> Self

pub fn activity_task_timed_out_event_attributes(
self,
input: ActivityTaskTimedOutEventAttributes
) -> Self

If the event is of type ActivityTaskTimedOut then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_activity_task_timed_out_event_attributes(
self,
input: Option<ActivityTaskTimedOutEventAttributes>
) -> Self

pub fn activity_task_canceled_event_attributes(
self,
input: ActivityTaskCanceledEventAttributes
) -> Self

If the event is of type ActivityTaskCanceled then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_activity_task_canceled_event_attributes(
self,
input: Option<ActivityTaskCanceledEventAttributes>
) -> Self

pub fn activity_task_cancel_requested_event_attributes(
self,
input: ActivityTaskCancelRequestedEventAttributes
) -> Self

If the event is of type ActivityTaskcancelRequested then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_activity_task_cancel_requested_event_attributes(
self,
input: Option<ActivityTaskCancelRequestedEventAttributes>
) -> Self

pub fn workflow_execution_signaled_event_attributes(
self,
input: WorkflowExecutionSignaledEventAttributes
) -> Self

If the event is of type WorkflowExecutionSignaled then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_workflow_execution_signaled_event_attributes(
self,
input: Option<WorkflowExecutionSignaledEventAttributes>
) -> Self

pub fn marker_recorded_event_attributes(
self,
input: MarkerRecordedEventAttributes
) -> Self

If the event is of type MarkerRecorded then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_marker_recorded_event_attributes(
self,
input: Option<MarkerRecordedEventAttributes>
) -> Self

pub fn record_marker_failed_event_attributes(
self,
input: RecordMarkerFailedEventAttributes
) -> Self

If the event is of type DecisionTaskFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_record_marker_failed_event_attributes(
self,
input: Option<RecordMarkerFailedEventAttributes>
) -> Self

pub fn timer_started_event_attributes(
self,
input: TimerStartedEventAttributes
) -> Self

If the event is of type TimerStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_timer_started_event_attributes(
self,
input: Option<TimerStartedEventAttributes>
) -> Self

pub fn timer_fired_event_attributes(
self,
input: TimerFiredEventAttributes
) -> Self

If the event is of type TimerFired then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_timer_fired_event_attributes(
self,
input: Option<TimerFiredEventAttributes>
) -> Self

pub fn timer_canceled_event_attributes(
self,
input: TimerCanceledEventAttributes
) -> Self

If the event is of type TimerCanceled then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_timer_canceled_event_attributes(
self,
input: Option<TimerCanceledEventAttributes>
) -> Self

pub fn start_child_workflow_execution_initiated_event_attributes(
self,
input: StartChildWorkflowExecutionInitiatedEventAttributes
) -> Self

If the event is of type StartChildWorkflowExecutionInitiated then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_start_child_workflow_execution_initiated_event_attributes(
self,
input: Option<StartChildWorkflowExecutionInitiatedEventAttributes>
) -> Self

pub fn child_workflow_execution_started_event_attributes(
self,
input: ChildWorkflowExecutionStartedEventAttributes
) -> Self

If the event is of type ChildWorkflowExecutionStarted then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_child_workflow_execution_started_event_attributes(
self,
input: Option<ChildWorkflowExecutionStartedEventAttributes>
) -> Self

pub fn child_workflow_execution_completed_event_attributes(
self,
input: ChildWorkflowExecutionCompletedEventAttributes
) -> Self

If the event is of type ChildWorkflowExecutionCompleted then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_child_workflow_execution_completed_event_attributes(
self,
input: Option<ChildWorkflowExecutionCompletedEventAttributes>
) -> Self

pub fn child_workflow_execution_failed_event_attributes(
self,
input: ChildWorkflowExecutionFailedEventAttributes
) -> Self

If the event is of type ChildWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_child_workflow_execution_failed_event_attributes(
self,
input: Option<ChildWorkflowExecutionFailedEventAttributes>
) -> Self

pub fn child_workflow_execution_timed_out_event_attributes(
self,
input: ChildWorkflowExecutionTimedOutEventAttributes
) -> Self

If the event is of type ChildWorkflowExecutionTimedOut then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_child_workflow_execution_timed_out_event_attributes(
self,
input: Option<ChildWorkflowExecutionTimedOutEventAttributes>
) -> Self

pub fn child_workflow_execution_canceled_event_attributes(
self,
input: ChildWorkflowExecutionCanceledEventAttributes
) -> Self

If the event is of type ChildWorkflowExecutionCanceled then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_child_workflow_execution_canceled_event_attributes(
self,
input: Option<ChildWorkflowExecutionCanceledEventAttributes>
) -> Self

pub fn child_workflow_execution_terminated_event_attributes(
self,
input: ChildWorkflowExecutionTerminatedEventAttributes
) -> Self

If the event is of type ChildWorkflowExecutionTerminated then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_child_workflow_execution_terminated_event_attributes(
self,
input: Option<ChildWorkflowExecutionTerminatedEventAttributes>
) -> Self

pub fn signal_external_workflow_execution_initiated_event_attributes(
self,
input: SignalExternalWorkflowExecutionInitiatedEventAttributes
) -> Self

If the event is of type SignalExternalWorkflowExecutionInitiated then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_signal_external_workflow_execution_initiated_event_attributes(
self,
input: Option<SignalExternalWorkflowExecutionInitiatedEventAttributes>
) -> Self

pub fn external_workflow_execution_signaled_event_attributes(
self,
input: ExternalWorkflowExecutionSignaledEventAttributes
) -> Self

If the event is of type ExternalWorkflowExecutionSignaled then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_external_workflow_execution_signaled_event_attributes(
self,
input: Option<ExternalWorkflowExecutionSignaledEventAttributes>
) -> Self

pub fn signal_external_workflow_execution_failed_event_attributes(
self,
input: SignalExternalWorkflowExecutionFailedEventAttributes
) -> Self

If the event is of type SignalExternalWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_signal_external_workflow_execution_failed_event_attributes(
self,
input: Option<SignalExternalWorkflowExecutionFailedEventAttributes>
) -> Self

pub fn external_workflow_execution_cancel_requested_event_attributes(
self,
input: ExternalWorkflowExecutionCancelRequestedEventAttributes
) -> Self

If the event is of type ExternalWorkflowExecutionCancelRequested then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_external_workflow_execution_cancel_requested_event_attributes(
self,
input: Option<ExternalWorkflowExecutionCancelRequestedEventAttributes>
) -> Self

pub fn request_cancel_external_workflow_execution_initiated_event_attributes(
self,
input: RequestCancelExternalWorkflowExecutionInitiatedEventAttributes
) -> Self

If the event is of type RequestCancelExternalWorkflowExecutionInitiated then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_request_cancel_external_workflow_execution_initiated_event_attributes(
self,
input: Option<RequestCancelExternalWorkflowExecutionInitiatedEventAttributes>
) -> Self

pub fn request_cancel_external_workflow_execution_failed_event_attributes(
self,
input: RequestCancelExternalWorkflowExecutionFailedEventAttributes
) -> Self

If the event is of type RequestCancelExternalWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_request_cancel_external_workflow_execution_failed_event_attributes(
self,
input: Option<RequestCancelExternalWorkflowExecutionFailedEventAttributes>
) -> Self

pub fn schedule_activity_task_failed_event_attributes(
self,
input: ScheduleActivityTaskFailedEventAttributes
) -> Self

If the event is of type ScheduleActivityTaskFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_schedule_activity_task_failed_event_attributes(
self,
input: Option<ScheduleActivityTaskFailedEventAttributes>
) -> Self

pub fn request_cancel_activity_task_failed_event_attributes(
self,
input: RequestCancelActivityTaskFailedEventAttributes
) -> Self

If the event is of type RequestCancelActivityTaskFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_request_cancel_activity_task_failed_event_attributes(
self,
input: Option<RequestCancelActivityTaskFailedEventAttributes>
) -> Self

pub fn start_timer_failed_event_attributes(
self,
input: StartTimerFailedEventAttributes
) -> Self

If the event is of type StartTimerFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_start_timer_failed_event_attributes(
self,
input: Option<StartTimerFailedEventAttributes>
) -> Self

pub fn cancel_timer_failed_event_attributes(
self,
input: CancelTimerFailedEventAttributes
) -> Self

If the event is of type CancelTimerFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_cancel_timer_failed_event_attributes(
self,
input: Option<CancelTimerFailedEventAttributes>
) -> Self

pub fn start_child_workflow_execution_failed_event_attributes(
self,
input: StartChildWorkflowExecutionFailedEventAttributes
) -> Self

If the event is of type StartChildWorkflowExecutionFailed then this member is set and provides detailed information about the event. It isn't set for other event types.

pub fn set_start_child_workflow_execution_failed_event_attributes(
self,
input: Option<StartChildWorkflowExecutionFailedEventAttributes>
) -> Self

pub fn lambda_function_scheduled_event_attributes(
self,
input: LambdaFunctionScheduledEventAttributes
) -> Self

Provides the details of the LambdaFunctionScheduled event. It isn't set for other event types.

pub fn set_lambda_function_scheduled_event_attributes(
self,
input: Option<LambdaFunctionScheduledEventAttributes>
) -> Self

pub fn lambda_function_started_event_attributes(
self,
input: LambdaFunctionStartedEventAttributes
) -> Self

Provides the details of the LambdaFunctionStarted event. It isn't set for other event types.

pub fn set_lambda_function_started_event_attributes(
self,
input: Option<LambdaFunctionStartedEventAttributes>
) -> Self

pub fn lambda_function_completed_event_attributes(
self,
input: LambdaFunctionCompletedEventAttributes
) -> Self

Provides the details of the LambdaFunctionCompleted event. It isn't set for other event types.

pub fn set_lambda_function_completed_event_attributes(
self,
input: Option<LambdaFunctionCompletedEventAttributes>
) -> Self

pub fn lambda_function_failed_event_attributes(
self,
input: LambdaFunctionFailedEventAttributes
) -> Self

Provides the details of the LambdaFunctionFailed event. It isn't set for other event types.

pub fn set_lambda_function_failed_event_attributes(
self,
input: Option<LambdaFunctionFailedEventAttributes>
) -> Self

pub fn lambda_function_timed_out_event_attributes(
self,
input: LambdaFunctionTimedOutEventAttributes
) -> Self

Provides the details of the LambdaFunctionTimedOut event. It isn't set for other event types.

pub fn set_lambda_function_timed_out_event_attributes(
self,
input: Option<LambdaFunctionTimedOutEventAttributes>
) -> Self

pub fn schedule_lambda_function_failed_event_attributes(
self,
input: ScheduleLambdaFunctionFailedEventAttributes
) -> Self

Provides the details of the ScheduleLambdaFunctionFailed event. It isn't set for other event types.

pub fn set_schedule_lambda_function_failed_event_attributes(
self,
input: Option<ScheduleLambdaFunctionFailedEventAttributes>
) -> Self

pub fn start_lambda_function_failed_event_attributes(
self,
input: StartLambdaFunctionFailedEventAttributes
) -> Self

Provides the details of the StartLambdaFunctionFailed event. It isn't set for other event types.

pub fn set_start_lambda_function_failed_event_attributes(
self,
input: Option<StartLambdaFunctionFailedEventAttributes>
) -> Self

pub fn build(self) -> HistoryEvent

Consumes the builder and constructs a HistoryEvent