Module aws_sdk_swf::operation

Structs

CountClosedWorkflowExecutions

Returns the number of closed workflow executions within the given domain that meet the

CountOpenWorkflowExecutions

Returns the number of open workflow executions within the given domain that meet the

CountPendingActivityTasks

Returns the estimated number of activity tasks in the specified task list. The count

CountPendingDecisionTasks

Returns the estimated number of decision tasks in the specified task list. The count

DeprecateActivityType

Deprecates the specified activity type. After an activity type has

DeprecateDomain

Deprecates the specified domain. After a domain has been deprecated it cannot be used

DeprecateWorkflowType

Deprecates the specified workflow type. After a workflow type has

DescribeActivityType

Returns information about the specified activity type. This includes configuration

DescribeDomain

Returns information about the specified domain, including description and

DescribeWorkflowExecution

Returns information about the specified workflow execution including its type and some

DescribeWorkflowType

Returns information about the specified workflow type. This

GetWorkflowExecutionHistory

Returns the history of the specified workflow execution. The results may be split into

ListActivityTypes

Returns information about all activities registered in the specified domain that match

ListClosedWorkflowExecutions

Returns a list of closed workflow executions in the specified domain that meet the

ListDomains

Returns the list of domains registered in the account. The results may be split into

ListOpenWorkflowExecutions

Returns a list of open workflow executions in the specified domain that meet the

ListTagsForResource

List tags for a given domain.

ListWorkflowTypes

Returns information about workflow types in the specified domain. The results may be

PollForActivityTask

Used by workers to get an ActivityTask from the specified activity

PollForDecisionTask

Used by deciders to get a DecisionTask from the specified decision

RecordActivityTaskHeartbeat

Used by activity workers to report to the service that the ActivityTask represented by the specified taskToken is still making progress. The worker

RegisterActivityType

Registers a new activity type along with its configuration

RegisterDomain

Registers a new domain.

RegisterWorkflowType

Registers a new workflow type and its configuration settings in

RequestCancelWorkflowExecution

Records a WorkflowExecutionCancelRequested event in the currently running

RespondActivityTaskCanceled

Used by workers to tell the service that the ActivityTask identified

RespondActivityTaskCompleted

Used by workers to tell the service that the ActivityTask identified

RespondActivityTaskFailed

Used by workers to tell the service that the ActivityTask identified

RespondDecisionTaskCompleted

Used by deciders to tell the service that the DecisionTask identified

SignalWorkflowExecution

Records a WorkflowExecutionSignaled event in the workflow execution

StartWorkflowExecution

Starts an execution of the workflow type in the specified domain using the provided

TagResource

Add a tag to a Amazon SWF domain.

TerminateWorkflowExecution

Records a WorkflowExecutionTerminated event and forces closure of the

UndeprecateActivityType

Undeprecates a previously deprecated activity type. After an activity type has

UndeprecateDomain

Undeprecates a previously deprecated domain. After a domain has been undeprecated it can be used

UndeprecateWorkflowType

Undeprecates a previously deprecated workflow type. After a workflow type has

UntagResource

Remove a tag from a Amazon SWF domain.