Module aws_sdk_swf::output

Modules

count_closed_workflow_executions_output

See CountClosedWorkflowExecutionsOutput

count_open_workflow_executions_output

See CountOpenWorkflowExecutionsOutput

count_pending_activity_tasks_output

See CountPendingActivityTasksOutput

count_pending_decision_tasks_output

See CountPendingDecisionTasksOutput

deprecate_activity_type_output

See DeprecateActivityTypeOutput

deprecate_domain_output

See DeprecateDomainOutput

deprecate_workflow_type_output

See DeprecateWorkflowTypeOutput

describe_activity_type_output

See DescribeActivityTypeOutput

describe_domain_output

See DescribeDomainOutput

describe_workflow_execution_output

See DescribeWorkflowExecutionOutput

describe_workflow_type_output

See DescribeWorkflowTypeOutput

get_workflow_execution_history_output

See GetWorkflowExecutionHistoryOutput

list_activity_types_output

See ListActivityTypesOutput

list_closed_workflow_executions_output

See ListClosedWorkflowExecutionsOutput

list_domains_output

See ListDomainsOutput

list_open_workflow_executions_output

See ListOpenWorkflowExecutionsOutput

list_tags_for_resource_output

See ListTagsForResourceOutput

list_workflow_types_output

See ListWorkflowTypesOutput

poll_for_activity_task_output

See PollForActivityTaskOutput

poll_for_decision_task_output

See PollForDecisionTaskOutput

record_activity_task_heartbeat_output

See RecordActivityTaskHeartbeatOutput

register_activity_type_output

See RegisterActivityTypeOutput

register_domain_output

See RegisterDomainOutput

register_workflow_type_output

See RegisterWorkflowTypeOutput

request_cancel_workflow_execution_output

See RequestCancelWorkflowExecutionOutput

respond_activity_task_canceled_output

See RespondActivityTaskCanceledOutput

respond_activity_task_completed_output

See RespondActivityTaskCompletedOutput

respond_activity_task_failed_output

See RespondActivityTaskFailedOutput

respond_decision_task_completed_output

See RespondDecisionTaskCompletedOutput

signal_workflow_execution_output

See SignalWorkflowExecutionOutput

start_workflow_execution_output

See StartWorkflowExecutionOutput

tag_resource_output

See TagResourceOutput

terminate_workflow_execution_output

See TerminateWorkflowExecutionOutput

undeprecate_activity_type_output

See UndeprecateActivityTypeOutput

undeprecate_domain_output

See UndeprecateDomainOutput

undeprecate_workflow_type_output

See UndeprecateWorkflowTypeOutput

untag_resource_output

See UntagResourceOutput

Structs

CountClosedWorkflowExecutionsOutput

Contains the count of workflow executions returned from CountOpenWorkflowExecutions or

CountOpenWorkflowExecutionsOutput

Contains the count of workflow executions returned from CountOpenWorkflowExecutions or

CountPendingActivityTasksOutput

Contains the count of tasks in a task list.

CountPendingDecisionTasksOutput

Contains the count of tasks in a task list.

DeprecateActivityTypeOutput

DeprecateDomainOutput

DeprecateWorkflowTypeOutput

DescribeActivityTypeOutput

Detailed information about an activity type.

DescribeDomainOutput

Contains details of a domain.

DescribeWorkflowExecutionOutput

Contains details about a workflow execution.

DescribeWorkflowTypeOutput

Contains details about a workflow type.

GetWorkflowExecutionHistoryOutput

Paginated representation of a workflow history for a workflow execution. This is the up to date, complete and authoritative record of the events related to all tasks and events in the life of the workflow execution.

ListActivityTypesOutput

Contains a paginated list of activity type information structures.

ListClosedWorkflowExecutionsOutput

Contains a paginated list of information about workflow executions.

ListDomainsOutput

Contains a paginated collection of DomainInfo structures.

ListOpenWorkflowExecutionsOutput

Contains a paginated list of information about workflow executions.

ListTagsForResourceOutput

ListWorkflowTypesOutput

Contains a paginated list of information structures about workflow types.

PollForActivityTaskOutput

Unit of work sent to an activity worker.

PollForDecisionTaskOutput

A structure that represents a decision task. Decision tasks are sent to deciders in order for them to make decisions.

RecordActivityTaskHeartbeatOutput

Status information about an activity task.

RegisterActivityTypeOutput

RegisterDomainOutput

RegisterWorkflowTypeOutput

RequestCancelWorkflowExecutionOutput

RespondActivityTaskCanceledOutput

RespondActivityTaskCompletedOutput

RespondActivityTaskFailedOutput

RespondDecisionTaskCompletedOutput

SignalWorkflowExecutionOutput

StartWorkflowExecutionOutput

Specifies the runId of a workflow execution.

TagResourceOutput

TerminateWorkflowExecutionOutput

UndeprecateActivityTypeOutput

UndeprecateDomainOutput

UndeprecateWorkflowTypeOutput

UntagResourceOutput